differentiate build and test config
