--- dash-renderer/jest.init.js
@@ -1,1 +1,0 @@
-import '@babel/polyfill';
