+++ dash-renderer/webpack.test.config.js
@@ -1,0 +1,8 @@
+const NodePolyfillPlugin = require('node-polyfill-webpack-plugin');
+const config = require('./webpack.base.config');
+
+module.exports = config({
+    plugins: [
+        new NodePolyfillPlugin()
+    ]
+});
